split init into separate files
