+++ client/hawc_client/base_client.py
@@ -1,0 +1,29 @@
+from io import StringIO
+
+import pandas as pd
+
+from .hawc_session import HawcSession
+
+
+class BaseClient:
+    """
+    Base client class.
+
+    Initiates with a given HawcSession object.
+    """
+
+    def __init__(self, session: HawcSession):
+        self.session = session
+
+    def _csv_to_df(self, csv: str) -> pd.DataFrame:
+        """
+        Takes a CSV string and returns the pandas DataFrame representation of it.
+
+        Args:
+            csv (str): CSV string
+
+        Returns:
+            pd.DataFrame: DataFrame from CSV
+        """
+        csv_io = StringIO(csv)
+        return pd.read_csv(csv_io)
